+++ source/components/window/window.js
@@ -265,6 +265,12 @@
                         .attr("tabindex", -1)
                         .html(item.html);
 
+                    if (item.attr && typeof item.attr === 'object') {
+                        $.each(item.attr, function(k, v){
+                            customButton.attr($.dashedName(k), v);
+                        });
+                    }
+
                     customButton.data("action", item.onclick);
 
                     buttons.prepend(customButton);